Merged PR 314: Merge review/fix-errors-and-warnings to master
Fixed a few compiler warnings.

Added missing include files to build on clang

Added an option to build using clang in Linux

Fixed a problem where we were linking to extra LLVM libraries we didn't need, which only seemed to cause problems when building with clang on Linux, and when trying to use the node module on the Mac.