Merged PR 322: Merge review/mathv2 to master
This is the first of 3 PRs intended to clean up the math library.
The goal for this PR is to fix the loopy dependence between Operations and Vector/Matrix/Tensor.

(the second PR will be general cleanup and unification of Vector/Matrix/Tensor; the third PR will add missing Operations that I owe Byron)