Merged PR 324: Merge review/layer-nodes to master
Adds two neural network nodes: `BiasLayerNode` and `BatchNormalizationLayerNode`.

Both of them refine into a `BroadcastLinearFunctionNode`, which evaluates the function `f(x) = a*x + b`, where `x`can be multi-dimensional data (a tensor), and `a`, and `b` are vectors along one of the dimensions of `x`. The values of `a` and `b` are replicated along the perpendicular dimensions. If multiple `BroadcastLinearFunctionNode`s are chained together in a model, they will be coalesced during refinement.

Eventually, I think we should relax the shape constraint on `a` and `b`, and allow general "broadcast" semantics. Then we can replace `Unary`/`BinaryOperationNode` with the "broadcast" versions.