Merged PR 351: Fix profiler on Windows target machines. (Needed to implement clock_gettime)
Fix profiler on Windows target machines. (Needed to implement clock_gettime using QueryPerformanceCounter).  Turns out just adding clock_gettime to the test itself doesn't work, the jitter is not that smart, it has to be in the jitted module.  Generating the code in header file doesn't work for JIT scenarios, which we are allowing our customers to use, so this is really the only fix that works in all cases.