Merged PR 360: jazzed up the readmes
My intention with the main README is that only the current release would have a screen shot image. Once a release is no longer the most recent, its image is retired, but its short paragraph could be kept. Every tutorial should have a screen shot to act as an "exciter" for the user.