Merged PR 376: automatic detection and enforcement of llvm version 3.9
automatic detection and enforcement of llvm version 3.9

Previously the readme encouraged the installation of *any* llvm version, as of version 4, include paths changed for llvm meaning that the build was broken.

Another constraint was the static path to an arbitrary version of llvm.

This commit enables the "automatic" detection of llvm paths on MacOS using HINTS to where the binary files might be, and places a strict version dependency of 3.9. The mac read me was updated accordingly.