Merged PR 442: Merge review/merge_rob to master
This change merges some of the code Rob wrote for the Jupyter Tutorial, specifically a nice Downloader class and a set of ELL Build Tool wrappers for swig, llc and compile.exe.  It also renames make_python_project to "wrap" as discussed this morning.  It still only supports Python language for now.