Merged PR 496: Merge review/fold-linear-ops to master
Added support for the "foldLinearOperations" flag on `compile` and `profile`.

This PR changes the signature for `Node::Refine()`. It now takes a pointer to the compiler that called `Refine` on the model (or `nullptr` if `Refine` was called outside of compilation). We'll want to revisit this at some point to clean up how the compiler and model refinement interact.