Merged PR 504: Updated BinaryConvolutionalLayerNode to use 32-bit packed data on 32-bit systems
Now the compiler can choose the width of the integer type it packs the binarized data used by binary convolutional layers into. It uses 32-bit ints if the `targetDevice`'s `numBits` is 32, and a 64-bit in otherwise. There should perhaps be a more sophisticated method for choosing later (and it should probably be overridable, if only for testing).