Merged PR 510: fix IR emitter bug
There was previously a bug where bad identifiers were generated if the IR emitter was used more than once at run time. (We never saw this in our batch programs, like compile, but it's easy to repro in a long-lived Python scripting session.) This was due to a global variable in IRModuleEmitter holding symbol state between all runs. I created a new unit test TestTwoEmitsInOneSession that manifested the bug. I've also checked in a fix, which passes all unit tests including the new one. However, because I'm a C++ n00b, I may be leaking memory, so that's a thing to check for.