Merged PR 550: Added stricter warning settings for Clang/GCC
Added `-Wall`, `-Wno-missing-braces`, and `-Wmissing-field-initializers` flags when using GCC or Clang. The `-Wno-missing-braces` is there because clang emits ridiculous amounts of warning goop for missing braces in cases that are perfectly fine.

Added `/W4` option for MSVC.

Adding these settings causes a lot of warnings (especially in recent code I added to the nodes library) that will want to be cleaned up. I'd rather do the cleanup incrementally, though, and have this PR only add the compiler options.