Merged PR 628: Merge review/retarget to master
#837

Initial commit for the 'retarget' commandline tool, enabling single class binary classification (i.e. squirrel / not a squirrel). It takes a neural network, cuts it either at a layer or a node, uses that as a featurizer for a linear predictor which is then trained and the resulting map is saved as output. Wrap can then be used to compile the result which can be deployed to the pi.

Example usage:
* retarget.exe --neuralNetworkFilename dsb_I64x64x3CMCMCMCMCCCMC1AS.ell --removeLastLayers 4 --inputDataFilename dataset.gsdf --outputModelFilename model1.ell
* retarget.exe --neuralNetworkFilename dsb_I64x64x3CMCMCMCMCCCMC1AS.ell --targetNodeId 1109 --inputDataFilename dataset.gsdf --outputModelFilename model1.ell

The LinearPredictor is trained using SDCA and stops either at maxEpochs or when the duality gap is at the required precision. A warning is printed if maxEpochs is reached without the desired duality gap being achieved.

In addition to the retarget tool, this PR contains some other minor changes:
- Update print tool to include node ids in output. In a future PR, I'll update ELL View to do the same.
- Fix a minor error in emitters for vector dot product where a float value was being initialized with a double (error caused the resulting retargeted model to fail to compile)

Related work items: #837