Merged PR 650: Add property-bag-type metadata to Model, DynamicMap, Node, and Port objects, plus archiver improvements to support the metadata.
Adds a new "PropertyBag" class that represents a key-value store with strings as the keys and arrays of strings as the values.

Adds one of these property bag classes to Nodes, Ports, Models, and DynamicMaps.

Adds better support for dealing with version numbers when unarchiving files. Now the `IArchivable` interface includes a method (`CanReadVersion`) that a class can override to tell the archiver if it can read an object with a given archive version number. The default implementation is to only read objects archived with the current archive version number for the class.

Added a central place (an enum) to keep track of model version numbers.

The classes that have metadata use this to be able to read the current model files from the gallery.

Related work items: #1033, #1035