Merged PR 666: Merge review/mac-archive-fix to master
There was a tiny typo in the most recent pull request that caused the build to break on the Mac. The problem was in a particularly horrible part of a particularly horrible part of the codebase: the macros that implement some boilerplate code in the archivers. Instead of just fixing the typo, I tried to clean things up to make it more maintainable in the future.

I also used the same "x macro" technique to make it easier to keep the array of NN layer type names in sync with the layer type enum. Hopefully it's an improvement.