Merged PR 700: Added new interface for emitting control-flow constructs from `IRFunctionEmitter`
This PR adds a new (IMO, better) way to create loops and if/else constructs. It prevents a common error: forgetting to close loops and if statements.

In the old version, the loop-initiating function (`IRFunctionEmitter::ForLoop`) would return a special "emitter" object and enter a "in a loop" state (which must be exited by calling `End` on the emitter object, or else hard-to-debug crashes occur). Now, the function that initiates the loop (`IRFunctionEmitter::For`) accepts a lambda that defines the body of the loop.

Old:
```
auto loop = function.ForLoop();
loop.Begin(10);     // Note: I often forget this semicolon, I think because it's unnatural to follow a semicolon with a '{'
{
    auto index = loop.LoadIterationVariable();
    auto x = function.ValueAt(myPtr, index);
    ...
}
loop.End();  // This statement is easy to forget.
```

New:
```
function.For(10, [](emitters::IRFunctionEmitter& function, llvm::Value* index){
    auto x = function.ValueAt(myPtr, index);
    ...
});
```

If blocks are a little different in that they return a reference to an "if emitter" object, which is only intended to be used to chain together elseif / else blocks:

```
function.If(test1, [](emitters::IRFunctionEmitter& function){
  ...
}).ElseIf(test2, [](emitters::IRFunctionEmitter& function){
  ...
}).Else([](emitters::IRFunctionEmitter& function){
  ...
});
```

The loops and if blocks in `IRThreadPool.cpp` have been migrated to the new method, as an example.

To avoid thrashing everyone's code, it doesn't remove the old methods, or do a wholesale conversion from old to new in the codebase. I was going to have that be a follow-up work item. Also, I didn't want to have to make a zillion edits every time someone had a good suggestion for a change. :)

Related work items: #1130