Merged PR 697: Added `IRLocalValue` classes to make it easier to emit IR
## Description

This PR adds a simple class (`IRLocalValue`) that is used to represent a value local to a function being emitted. The class wraps an LLVM value and a reference to the `IRFunctionEmitter` for the function. A subclass, `IRLocalScalar` is used to represent scalar (non-pointer) values, and a number of helpful function and operator overrides are included to make it easier to do simple computation with these values.

[See first comment for fuller description -- I've gone over my allotted 4k characters]

Related work items: #1132