Merged PR 691: Add ccache support for faster build times
This PR adds some very simple support for using `ccache` to improve build times. The `ccache` tool intercepts the compiler and somehow does a much better job than `make` at determining if it can use an existing object file rather than rebuilding it. For me, it can result in a significant speedup in incremental rebuild time.

Using `ccache` is disabled by default. Enabling it on involves either setting the `USE_COMPILER_CACHE` CMake variable or the `CMAKE_COMPILER_CACHE` environment variable to "ON".

Sadly, it only works on Linux and the Mac. However, there is a similar thing for Windows called `clcache`. I've tried a little bit to get it to work, but haven't succeeded so far. If anyone else would like to experiment with getting it to work, you can check out the `dev/cjacobs/ccache` branch, which adds the `clcache` NuGet package to our list of package configs we rely on, and adds some CMake code to attempt to invoke it.

I do want to add `clcache` support soon, but would like to get this merged in now, since it improves build time for me so much.

Related work items: #1102