Merged PR 728: Clean up headers in utilities, emitters, and model libraries
This PR does a couple of things:

1) Remove unnecessary headers files that are being included
2) Forward-declare classes instead of including their header files when appropriate

Item 1 should be uncontroversial. Item 2 is a little squishy, though. If we are going to use forward declarations in our code, we may want to have a guideline on when it's appropriate to use them. The Internet isn't in any kind of sharp agreement about when forward defines are appropriate: Google's style guide says "never", Scott Meyers says "whenever possible". Maybe we can hash out a set of rules in the comments of this PR.

Here are a few thoughts that might eventually congeal into something like rules:

* Forward declarations only really make sense in header files.
* Forward declarations within the same library seem safest
* Forward declarations in other ELL libraries seem perhaps OK, but less safe. They're kind of ugly, since they require putting the declaration within a different namespace than the main namespace of the file.
* In 3rd-party libraries (say, LLVM) forward declarations seem somehow less safe than ones between ELL libraries.
* While it's legal to forward-declare an enum class, it's kind of ugly and can cause an error if the enum's underlying datatype changes. Maybe avoid them.
* According to internet people, forward declarations of templated classes are particularly dangerous. Maybe it's best to avoid them? On the other hand, they're likely to be the ones that give the most speedup to the build process.

---------

On my laptop, the changes in this PR sped up build times a little. Here are the times to build each of the libraries, plus a couple of the command-line tools. The time reported is the time to build the library in question from scratch, given that all the prerequisite libraries have been built. It's not necessarily a great test, but it was easy to automate. I stuck an 'X' at the end for row that seemed to maybe have helped a noticeable amount.

| library | before | after |.   |
|--------|---------|------|----|
| utilities | 0m16.393s | 0m16.383s |
| math | 0m1.228s | 0m1.369s |
| emitters | 0m45.885s | 0m32.001s | X
| data | 0m4.991s | 0m4.726s |
| dsp | 0m0.852s | 0m0.816s |
| functions | 0m1.939s | 0m1.832s |
| evaluators | 0m0.838s | 0m0.871s |
| predictors | 0m2.046s | 0m1.984s |
| trainers | 0m7.098s | 0m6.880s |
| model | 0m45.083s | 0m32.945s | X
| nodes | 1m36.911s | 1m19.710s | X
| common | 1m10.994s | 1m0.129s | X
| testing | 0m7.471s | 0m6.239s |
| compile | 0m9.213s | 0m6.356s | X
| apply | 0m6.895s | 0m5.633s |