Merged PR 790: Added ModelOptimizer class for (wait for it...) optimizing models
This PR has a simple optimizer class that performs transformations on the input model (the idea being to transform it into a model that, when compiled, is more efficient than it would be otherwise). The main new classes are:

`ModelOptimizer` --- A class that contains and invokes the different optimization passes.
`OptimizationPass` --- Base class for optimization passes.
`NodeLocalOptimizationPass` --- A subclass of `OptimizationPass` that performs local modifications to the model graph.
`OptimizationPassRegistry` --- A global registry of optimization passes. When initializing a `ModelOptimizer`, the map compiler adds passes from registry that are enabled by the compiler's options.
`FuseLinearOperationsPass` --- A concrete pass that combines multiple `BroadcastLinearFunctionNode`s that are connected together in the right way.

The core optimizer files are in the `model` library, but in a separate subdirectory. I'm not sure they need to be, though. If they stay where they are, maybe they ought to be in a sub-namespace?

I'm not sure I love the global registry of optimization passes. But I'm not sure I hate it either.

I didn't make many changes to `ModelTransformer`, not because it doesn't have it coming, but just to get this PR out the door.

I seem to have added a markdown file with a bunch of blathering to myself in it. Don't let me forget to remove it.

Related work items: #1282