Merged PR 788: Merge review/refactorCntkImporter to master
New version of importer. This is a staged change to separate out the code required for moving to Nodes.

The importer is now conceptually 3 parts:
- The front end importer. This reads in native format (e.g. CNTK) and creates an intermediate representation via ImporterNodes
- The ImporterEngine, which drives conversion process from ImporterNodes to a list of ELL components
- The back-end ELL converters, which do conversion of an individual ImporterNode to an individual ELL component. For now, the back end is using ELL neural network layers.
The converters in both the from and back ends are specified via table mappings.
Note that new importers don't need to touch ImporterEngine or the back end Converters, they only need to implement the front end import.

This change also contains:
- Infrastructure changes for the CNTK front end importer to parse arbitrary CNTK graphs (the backend does not support this yet - it will when we move to create Nodes instead of Layers). This was tested on Densely connected Convolutional Networks.
- For compatibility, the legacy importer is still run by default. The new importer is run with the "--use_importer_engine" switch. I'll switch it over once the move to Nodes is complete.

The new importer was tested on all models in the gallery.

Related work items: #1160