Merged PR 767: Added Winograd convolution node; added convolution functions to dsp library.
Adds convolution functions to the dsp library.

Adds new convolution nodes for "simple" and "Winograd" convolution types.

**Note:**
It might be best to examine the changes to the nodes library first. I'm not very happy with the code added to the dsp library, and it may see some changes yet.

Related work items: #1271, #1272