Merged PR 831: add a void* context parameter to predict that is passed through to the callback functions
add a void* context parameter to predict that is passed through to the callback functions.  This makes it easier to implement callbacks in object oriented code.  It also solves a problem with memory leak in Python callback wrappers since we now have a place to store the callback objects in the context which helps tie their lifetime to the context which means when the context goes away, so does the callback objects.

Related work items: #1194