Merged PR 855: Stop using assert to validate arguments given to public API's.
I was just debugging a “crash” in our compiler and noticed it was an assert, and so I’m wondering what we decided in terms of policy for using assert versus throw.  Seems our code base is inconsistent on this.   For example, the NeuralNetworkPredictor we have this in its constructor:

        assert(input.Size() == GetShapeSize(_predictor.GetInputShape()));

And MatrixMultiplyNode has this:

        if (input1.Size() != m * k)
        {
            throw utilities::InputException(utilities::InputExceptionErrors::invalidArgument, "Input matrix 1 size incorrect");
        }

I would argue that “crash” is a bad experience that should be avoided, so I recommend “assert” only be used to guard against bugs in our code and that anything else (like bad input from user who builds a model using our API) should be an invalidArgument exception.

We also have DEBUG_THROW so we need guidelines on when to use this one too.  Since it only throws in Debug compile of our bits, clearly DEBUG_THROW has to mean that it is ok to proceed without the throw in release bits.  I can’t actually think of a good example when this should be used.  Most places where I saw it being used would pretty much result in a crash in release code.  So I’m not convinced we should be using this one at all.

To this end I’ve gone through our code and produced this pull request to give an idea of the scope of this change.

Related work items: #1451