Merged PR 850: fix docs for opencv setup on raspberry pi.
Fix docs for opencv setup on raspberry pi, turns out jessie doesn't work if we label "stretch" version with the label "main".  conda is silly and picks the latest version even when you specify an explicit label.  So I removed the "main" label and added explicit label to our docs and added the missing addition of the Microsoft-ell channel which is needed to resolve the jasper and jpg packages.

Related work items: #1436