Merged PR 873: Add parallel for loops to emitters
This PR adds parallel for loops to the emitters library. The interface is similar to normal for loops, with the added requirement that local values need to be passed via a "capturedValues" argument (instead of being captured by the lambda representing the body of the loop. (This requirement is because these captured values need to be passed as arguments to the worker function executed by the thread pool).

The `SimpleConvolutionNode` has been parallelized via parallel-for. Here's what changed in the implementation:

Before:

```
        const auto numFilters = outputLayout.GetActiveSize(2);
        function.For(numFilters, [input, filterWeights, result, inputLayout, outputLayout, inputMemoryIncrements, outputMemoryIncrements, filterSize, stride](emitters::IRFunctionEmitter& function, llvm::Value* loopIndex1, const std::vector<llvm::Value*>& capturedValues) {
            // Loop body
        });
```

After:
```
        const auto numFilters = outputLayout.GetActiveSize(2);
        function.ParallelFor(numFilters, {input, filterWeights, result}, [inputLayout, outputLayout, inputMemoryIncrements, outputMemoryIncrements, filterSize, stride](emitters::IRFunctionEmitter& function, llvm::Value* loopIndex1, const std::vector<llvm::Value*>& capturedValues) {
            auto input = capturedValues[0];
            auto filterWeights = capturedValues[1];
            auto result = capturedValues[2];
            // Loop body
        });
```

Related work items: #1450