Depthwise separable convolutions use simple convolution by default This PR changes the looping order for separable depthwise simple convolutions to row-major order instead of channel-major order.
Additionally, it changes the behavior of `ConvolutionalLayer` to parse the value of `ConvolutionMethod` a little differently -- a value of `0` now means `automatic`, which allows the layer to select the method we think is most appropriate. Currently, this means unrolled convolutions for normal convolutions and, with this PR, simple convolutions for depthwise separable layers. Non-zero values are considered "forced" methods, which are adhered to unless the method is not possible with the rest of the parameters, in which case, the fallback method is unrolled.

While technically this changes the deserialization behavior, in practice, there's no change since for normal convolutions unrolled is the default and we haven't published depthwise separable models yet.

Note that this change doesn't do depthwise separable layers in channel-major order.

**Time for profiled models on Raspberry Pi Zero**
|                           | unrolled | simple  | simple + winograd |
|---------------------------|----------|---------|---------|
| separable big elderflower | 1286 ms  | 1086 ms | 1216 ms |
| separable brazilian guava | 197 ms   | 179 ms  | 194 ms |
| separable elderflower     | 246 ms   | 227 ms  | 249 ms |
| separable gardencress     | 1593 ms  | 1414 ms | 1563 ms |
| separable pandanflower    | 1741 ms  | 1730 ms | 1612 ms |

**Time for profiled models on Raspberry Pi 3**
|                           | unrolled | simple  | simple + winograd |
|---------------------------|----------|---------|---------|
| separable big elderflower | 247 ms  | 175 ms | 217 ms |
| separable brazilian guava | 42 ms   | 35 ms  | 38 ms |
| separable elderflower     | 56 ms   | 44 ms  | 50 ms |
| separable gardencress     | 301 ms  | 248 ms | 285 ms |
| separable pandanflower    | 319 ms  | 303 ms | 343 ms |

Story #1557: Get profile times of separable models on pi0
Story #1584: Fix ConvolutionalLayer always selection unrolled convolutions
Story #1514: Update DepthwiseFor to match reference with optimal ordering and parallelization