Merged PR 937: Fixes to vision demo scripts
The demo.py vision script doesn't work well on the "cifar" image test set which github customer is trying because the labels we draw on the image take over the entire image.  So I added a --print_labels option and also fixed the --nogui option so it works properly when running through test images with the --folder argument.