Merged PR 933: Add dimension order to MemoryLayout
Here's a simple start toward a way to add logical ordering of dimensions to the nodes and emitters parts of the codebase. In the current form, it basically just adds a permutation vector to go from "physical coordinates" to "logical coordinates" to the `MemoryLayout` class.

From here, there are a few obvious next steps:
* Add the obvious missing functionality to `MemoryLayout` (like a constructor that accepts a logical order).
* Change the various nodes that have a concept of ordering to use the one in `MemoryLayout`. I think these are:
    - ReorderDataNode
    - BinaryConvolutionalLayerNode
    - ConvolutionalLayerNode
    - WinogradConvolutionNode
    - ReceptiveFieldMatrixNode

* Use `MemoryLayout` inside `emitters::IRLocalMatrix`, `emitters::IRLocalTensor`, and `emitters::IRLocalMultidimArray`

* Migrate the orderings constants from `emitters::IRLocalMatrix` and `emitters::IRLocalTensor` to utilities

* Add functions for generating emitted code that efficiently iterates over memory. Something like
```
IRLocalArray bias;
IRLocalTensor memory;
function.For(memoryLayout, [memory, bias](auto function, auto indices, auto offset){
  auto biasValue = bias[indices[2]]; // channel = dimension 2
  memory[offset] *= biasValue;
});
```

Here, `indices` is a generalization of the loop iteration variable in for loops --- it's a vector of indices, one per dimension. And `offset` is a global offset (or iterator) into the memory. You can then efficiently index into the memory via `offset` if you don't want to do the math to compute the offset each time inside the loop. The last line in the above would be functionally identical to:
```
  memory[indices] *= biasValue;
```

I think we'd want to generalize the iterate-over-memory construct to loop in parallel over 2 or 3 chunks of memory as well, which could then be used to implement `copy` (which, with ordering present, would generalize to transpose as well) and the various binary operations.

If we used these constructs consistently, I think we could mark the generated code in such a way that it would be super-easy to fuse adjacent loops over the same dimensions. This would potentially make it simple to fuse activation function application into convolution.

Related work items: #1458