Merged PR 981: Remove Float/Double prefix from Neural layers in Python Interface
The ModelBuilder already uses a dtype parameter on the lower level node types which is the direction we want to go.  This change moves the Neural Layers and Predictor to the same model by moving the desired dataType to LayerParameters and removing the Double/Float prefix on types (meaning the Python interfaces for Layers and NeuralNetworkPredictor are no longer templates).