Merged PR 979: Fix problems with scalar port inputs
This PR fixes an inconsistency with how input ports are compiled. When an input port had a size of 1, then the value used to be passed to the node as a scalar, instead of a pointer to a chunk of memory with 1 element. This meant special-case code to make node `Compile` implementations work with scalar inputs.