Merged PR 992: Remove PortElements from InputPort
This PR removes the `PortElements` member of `InputNode` and replaces it with just a reference to an output port. Now, "complex" inputs (a subset of an output port, the concatenation of output ports, or some combination thereof) are explicitly generated as nodes in the model.

One part of the implementation is maybe a little gross: there's a (variadic template) function that intercepts the arguments passed to `Model::AddNode`, and if any of them are non-simple `PortElements`, extra nodes are added to compute the desired values. This allows the existing code that adds nodes to work without modification.

I haven't changed `ModelTransformer` much, but I think it can be simplified now that all of the input ports in the source model are simpler.

Related work items: #1686, #1687, #1688, #1689