Merged PR 1009: Remove most instances of PortElements from the codebase
This PR mostly does the straightforward cleanup implied by the previous one. Now that an `InputPort` refers to a single `OutputPort` as its source (instead of a complex `PortElements`), most of the instances of `PortElements` can be replaced by a reference to an `OutputPort` instead.

The "old->new" mapping part of `ModelTransformer` class is much simpler now, too, since it only has to deal with simple 1-1 port->port maps instead of potentially-complex many-many mappings.