Merged PR 1025: cleanup copy_newer_files pattern in our cmake files.
I found the problem that keeps causing cmake to rerun it's generation phase. The problem seems to be with 'configure_file' commands (at least on Windows) where changing any one of those files causes cmake to re-run it's very slow "generation" phase.  It is better to use a new macro "copy_newer_files" which uses a custom target with a copy command to copy those files.  This way cmake regeneration does not happen because the copy happens at build time instead of cmake generation time.  For example, the cntk importer CMakeLists.txt contains this:
```cmake
    foreach(PY_FILE ${importer_src})
        configure_file(${PY_FILE} ${PY_FILE} COPYONLY)
    endforeach()
```
and I would change it to this:
```cmake
    copy_newer_files(${module_name} ${importer_src})
```

Related work items: #1657