Merged PR 1043: Initial version of the value library for code nodes
This change adds the value library, which seeks to serve as the basis for the unification of the compute and compile versions of algorithms in our code.

It current has support for Scalar and Vector types, for-loop creations, and for summation, subtraction, multiplication, and division. This support is available for both normal "compute" mode as well as for emitting to LLVM IR, referred to as "compile mode".

Related work items: #1711, #1714, #1854