Merged PR 1056: Fix error in If emitter (and clean up emitters tests a bit)
There was a tricky issue with chaining `ElseIf` clauses to an emitted `If` statement. Sometimes the code to compute the condition was getting emitted into the "then" clause of the original if. That is, in the following pseudocode:

```
f.If(condition1, [](){ <then block 1> }).ElseIf(condition2, [](){ <then block 2> });
```

the code to compute condition2 was being emitted into `<then block 1>`, causing all sorts of trouble.

Additionally, I converted a bunch of the emitters tests that just print stuff into tests that verify stuff.

Related work items: #1875