Merged PR 1081: Fix swig build error on VS 15.9.2 with cmake version 3.12
Fix swig build error on VS 15.9.2 with cmake version 3.12.  The new version of cmake doesn't seem to use SWIG_LIB any more, instead it uses SWIG_DIR.  But it expects to find the libs in there, so SWIG_DIR now has to be equal to SWIG_LIB... now sure how to make it both ways compatible... but the build seems to pass so looks like we're good.