Merged PR 1076: Add `emittable_functions` library + tests; fix `value` library bugs
This change adds the beginnings of the `emittable_functions` library, which depends on the `value` library to provide function definitions for common operations that we're interested in.

Additionally, it includes the fixes in the `value` library to enable these functions. The most prominent change is that the `LLVMConte`xt promotes constant data to emittable data if the operation itself is emittable. An example is iterating over constant data `Vector` with an emittable index. In this case, the constant data `Vector` will be "promoted" to emittable, so that all future uses of the `Vector` now result in the emittable version of the data being used.

Another fix is that the `ComputeContext` wasn't properly handling the returning of global values.

Yet another change is the `Cast` function, which has been made a lot more user-friendly in its usage.

Related work items: #1894, #1905