Merged PR 1102: Revamp function creation, add support for calling intrinsics and external functions
This change removes the `CreateFunction` set of functions. In its place is now the `DeclareFunction` function, which allows for more explicit and easier to read function signature declaration. This allows simplified the `EmitterContext` interface.

Functions can now also have their names mangled, which is enabled by default, but can be disabled. It is specially important to disable name mangling when calling an externally defined function.

A number of intrinsics are defined, which are functions that are "built-in" to the library. These are:
* `Abs`
* `Cos`
* `Exp`
* `Log`
* `Max`
* `Min`
* `Pow`
* `Sin`
* `Sqrt`
* `Tanh`

It is now also possible to emit calls to externally defined functions for the `LLVMContext`. There is no such functionality in the `ComputeContext` since it doesn't make sense to do so.

The implementation of `Dot` has been updated to call directly into BLAS for both `ComputeContext` and `LLVMContext` as an example.

Related work items: #1856, #1857, #1881, #1886