Merged PR 1101: New functions for adding nodes
This PR adds some helper functions for adding nodes to models. These functions return the new node's output, instead of the node itself. It turns out, we almost always want to use a node's output, and almost never care about the function object itself. This change makes code that adds nodes to the model much nicer, and usually results in better variable names (as the variable now represents the quantity you want, not the function that computes it).

The naming convention for the new functions is `Append` + the node type (minus the word "Node"). So, the function to add a `SumNode`, is `AppendSum`. Each of these "append" functions has an overload that takes a `Model`, and one that takes a `ModelTransformer`, so you can use them when creating a model from scratch, and when refining a node.

Here's an example of client code using the old vs. new method of adding nodes:

```
        auto multNode = transformer.AddNode<BinaryOperationNode<ValueType>>(input1, input2, emitters::BinaryOperationType::coordinatewiseMultiply);
        auto sumNode = transformer.AddNode<SumNode<ValueType>>(multNode->output);
```

```
        const auto& product = AppendBinaryOperation(transformer, input1, input2, emitters::BinaryOperationType::coordinatewiseMultiply);
        const auto& sum = AppendSum(transformer, product);
```

I'm thinking that it might make more sense to remove the `Append` as well, at least when the node is named for an operation. So, the function to add a `SumNode` would be `Sum`, the function to add a `ReorderNode` would be `Reorder`, and so on. Additionally, adding helper functions for specific operations would make the code much nicer as well:

```
        const auto& product = CoordinatewiseMutiply(transformer, input1, input2);
        const auto& sum = Sum(transformer, product);
```

So far I've only added these helper functions for a few classes. I was going to wait for feedback before adding more.

--chuck