Merged PR 1141: Improve logging in ELL pythonlibs so it works better with ELL-training scripts.
Improve logger.py the same way I did in ELL-Training repo so that these loggers work better together. Specifically "--verbose" should instead be a logging verbosity level, which it now is.  Turn off terminal colors in runtest.sh.in so that the log doesn't contain weird characters.  Pickup RPI_* environment variables so pitest is enabled in that case.