Merged PR 1152: Make InputPort<T> and OutputPort<T> fully type erased
`InputPort<T>` and `OutputPort<T>` now no longer have any data members of their own and instead act as convenient wrappers to `InputPortBase` and `OutputPortBase`, respectively.  In the case of `InputPortBase`/`InputPort<T>`, this didn't require much of a change, since the base class already held all the data, which in this case is a just a pointer to an `OutputPortBase`. For `OutputPortBase`/`OutputPort<T>`, however, the templated data member has been moved to `OutputPortBase` and is powered by `std::variant`. This has the limitation that we must know all final concrete data types we make use of, but in practice, this isn't a problem as we have much more constraining limits elsewhere in the code.

A concern of slicing has been raised when it comes to this effort. To alleviate concerns, `static_assert` statements have been added to ensure that the layout of the base and derived classes never change.

Related work items: #1860, #2007