Merged PR 1100: Initial global optimizer PR
This is the initial PR for the global optimizer work. It replaces the original `ModelOptimizer` / `OptimizationPass` infrastructure with one using `Transformation` objects. The concept of refining the model has also been moved into a `Transformation`.

Most of the code having to do autotuning is is still a work-in-progress, and is relegated to the test directory so that it doesn't infect the library itself.

## Library code

In the `model` library, there are a few new classes:

**Transformation**: an abstract base class for transformations of the model. These transformations can be simple low-level transformations (e.g., set the algorithm to be used for a convolutional layer node), or can be complex (e.g., the global optimizer itself).

**RefineTransformation**: this is a `Transformation` subclass that refines a model. To refine a model, you now apply a `RefineTransformation` instead of calling `ModelTransformer::RefineModel`.

**OptimizeModelTransformation**: this is a `Transformation` subclass that runs the old `ModelOptimizer` on a model. It'll eventually go away when the passes that are part of the `ModelOptimizer` have been converted into `Transformation` subclasses.

**OptimizeAndRefineTransformation**: this is a `Transformation` subclass that encapsulates the series of "refine" and "optimize" transformations the compiler has to do before compiling the model. It'll eventually get turned into something more sane when the old `ModelOptimizer` has been retired.

There are a few other miscellaneous changes:
* `TransformContext` has been moved to its own file.
* As mentioned above, there is no `ModelTransformer::RefineModel` anymore. (I think we should get rid of `ModelTransformer::CopyModel` as well, and just have a `DeepCopy` method on `Model'.)
* `Model::Refine()` was removed from the python interface. Nobody used it as far as I could tell.

## Test code

There are a bunch of tests for the above-mentioned transformations, as well as a the implementation and tests of a number of "global optimizer" classes we'll need in the future, but don't rely on:
* `Cost`
* `CostModel`
* `Objective`
* ...

Related work items: #1733